requirements.txt
google-genai
flask
flask_socketio
flask_cors